chore: publish test utils and reorganize package
