+++ packages/test-utils-mcp/src/examples/hello/README.md
@@ -1,0 +1,3 @@
+# Hello MCP Server Example
+
+This example server exposes a single `sayHello` tool so tests can confirm that Model Context Protocol clients establish a session and invoke tools over Server-Sent Events and streamable HTTP transports. It does **not** exercise session state or storage; it is only intended to validate basic transport connectivity.
